refactor(admin): migrate collaborators-tags tab to canonical /social/... path
